requirements.txt
telebot
python-telegram-bot
pymongo
python-docx
pymupdf 
Pillow 